Rename the opencascade Manifold classes.
Manifolds have names ending in 'Manifold': now that Boundary is gone these names
do not make sense.

For backwards compatibility, the old classes now simply inherit from the
correctly-named Manifold versions. This requires marking some things that were
private now as protected.
