+++ source/opencascade/manifold_lib.inst.in
@@ -17,10 +17,14 @@
 
 for (deal_II_dimension : DIMENSIONS)
 {
-    template class NormalProjectionBoundary<deal_II_dimension, 3>;
-    template class DirectionalProjectionBoundary<deal_II_dimension, 3>;
-    template class NormalToMeshProjectionBoundary<deal_II_dimension, 3>;
+    template class NormalProjectionManifold<deal_II_dimension, 3>;
+    template class DirectionalProjectionManifold<deal_II_dimension, 3>;
+    template class NormalToMeshProjectionManifold<deal_II_dimension, 3>;
+    template class ArclengthProjectionLineManifold<deal_II_dimension, 3>;
+    template class NURBSPatchManifold<deal_II_dimension, 3>;
 #if deal_II_dimension <= 2
-    template class DirectionalProjectionBoundary<deal_II_dimension, 2>;
+    template class DirectionalProjectionManifold<deal_II_dimension, 2>;
+    template class ArclengthProjectionLineManifold<deal_II_dimension, 2>;
+    template class NURBSPatchManifold<deal_II_dimension, 2>;
 #endif
 }
